fix client placeholders
