+++ src/feed/AprscFeed.h
@@ -75,7 +75,7 @@ public:
      */
     ~AprscFeed() noexcept;
 
-    void registerToClient(client::ClientManager& rManager) override;
+    void registerToClient(client::ClientManager<client::ConnectorImplBoost>& rManager) override;
 
     /**
      * @see Feed#process
