+++ src/feed/AtmosphereFeed.cpp
@@ -26,7 +26,7 @@
 #include "../config/Configuration.h"
 #include "../data/AtmosphereData.h"
 #include "../object/Atmosphere.h"
-#include "client/ClientManager.h"
+#include "client/ClientManager.hpp"
 #include "parser/AtmosphereParser.h"
 
 namespace feed
@@ -41,11 +41,11 @@ AtmosphereFeed::AtmosphereFeed(const std::string& crName, const config::KeyValue
 AtmosphereFeed::~AtmosphereFeed() noexcept
 {}
 
-void AtmosphereFeed::registerToClient(client::ClientManager& rManager)
+void AtmosphereFeed::registerToClient(client::ClientManager<client::ConnectorImplBoost>& rManager)
 {
-    rManager.subscribe(
-        shared_from_this(), {mKvMap.find(KV_KEY_HOST)->second, mKvMap.find(KV_KEY_PORT)->second},
-        client::ClientManager::Protocol::SENSOR);
+    rManager.subscribe(shared_from_this(),
+                       {mKvMap.find(KV_KEY_HOST)->second, mKvMap.find(KV_KEY_PORT)->second},
+                       client::ClientManager<client::ConnectorImplBoost>::Protocol::SENSOR);
 }
 
 bool AtmosphereFeed::process(const std::string& crResponse) noexcept
