+++ src/feed/AtmosphereFeed.h
@@ -73,7 +73,7 @@ public:
      */
     ~AtmosphereFeed() noexcept;
 
-    void registerToClient(client::ClientManager& rManager) override;
+    void registerToClient(client::ClientManager<client::ConnectorImplBoost>& rManager) override;
 
     /**
      * @see Feed#process
