+++ src/feed/SbsFeed.h
@@ -74,7 +74,7 @@ public:
      */
     ~SbsFeed() noexcept;
 
-    void registerToClient(client::ClientManager& rManager) override;
+    void registerToClient(client::ClientManager<client::ConnectorImplBoost>& rManager) override;
 
     /**
      * @see Feed#process
