+++ src/feed/WindFeed.h
@@ -73,7 +73,7 @@ public:
      */
     ~WindFeed() noexcept;
 
-    void registerToClient(client::ClientManager& rManager) override;
+    void registerToClient(client::ClientManager<client::ConnectorImplBoost>& rManager) override;
 
     /**
      * @see Feed#process
